Bs12 with tgport mergefixes

+++ code/modules/admin/admin_verbs.dm
@@ -111,12 +111,12 @@
 			return
 
 		//Admin Observer
-		if (holder.level >= -1)
+		if (holder.level == -1)
 			verbs += /client/proc/investigate_show
 			verbs += /client/proc/cmd_admin_say
 			verbs += /client/proc/cmd_admin_gib_self
 			verbs += /client/proc/deadmin_self
-		else	return
+			return
 
 		//Moderator
 		if (holder.level >= 0)
@@ -127,12 +127,18 @@
 			verbs += /client/proc/Report
 			verbs += /client/proc/display_admin_reports
 			verbs += /datum/admins/proc/show_skills
-			verbs += /client/proc/mod_panel
 			verbs += /client/proc/admin_ghost
 		else	return
 
+		//Extra moderator commands
+		if(holder.level == 0)
+			verbs -= /client/proc/cmd_admin_say
+			verbs += /client/proc/mod_panel
+			return
+
 		//Temporary Admin
 		if (holder.level >= 1)
+			verbs += /client/proc/investigate_show
 			verbs += /datum/admins/proc/startnow
 			verbs += /datum/admins/proc/delay								//game start delay
 			verbs += /datum/admins/proc/immreboot							//immediate reboot
@@ -142,6 +148,8 @@
 			verbs += /client/proc/toggle_hear_deadcast
 			verbs += /client/proc/toggle_hear_radio
 			verbs += /client/proc/deadmin_self
+			verbs += /client/proc/player_panel_new
+			verbs += /client/proc/cmd_admin_change_custom_event
 			//verbs += /client/proc/cmd_admin_attack_log				--Merged with view variables
 		else	return
 
@@ -275,6 +283,7 @@
 			verbs += /client/proc/startSinglo							//Used to prevent the station from losing power while testing stuff out.
 			verbs += /client/proc/toggle_log_hrefs
 			verbs += /client/proc/cmd_debug_mob_lists
+			verbs += /client/proc/set_ooc
 		else	return
 	return
 
@@ -404,6 +413,7 @@
 		/datum/admins/proc/access_news_network,
 		/client/proc/one_click_antag,
 		/client/proc/invisimin,
+		/client/proc/set_ooc,
 
 		//bs12 verbs
 		/client/proc/update_mob_sprite,
@@ -415,7 +425,8 @@
 		/client/proc/display_admin_reports,
 		/client/proc/editappear,
 		/client/proc/cmd_mod_say,
-		/client/proc/playernotes
+		/client/proc/playernotes,
+		/client/proc/cmd_admin_change_custom_event
 	)
 	//verbs -= /client/proc/mapload
 	//verbs -= /client/proc/cmd_admin_drop_everything					--merged with view variables
